Agentify_Backend/requirements.txt
pydantic~=2.9.2
python-dotenv~=1.0.1
langchain~=0.3.7
joblib~=1.4.2
pandas~=2.2.3
scikit-learn~=1.3.2
requests~=2.32.3
fastapi
uvicorn
langchain-google-genai
sentence-transformers
langchain_huggingface
langchain_community
pinecone
pinecone-client